pyproject.toml
[project]
name = "agents-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "alembic>=1.17.0",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.119.0",
    "passlib[bcrypt]>=1.7.4",
    "pyautogen>=0.10.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "anyio>=4.11.0",
    "httpx>=0.28.1",
    "pyrefly>=0.36.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.14.0",
]
